+++ Editor/Camera.h
@@ -11,13 +11,13 @@ namespace UltraEd
         Camera();
         Camera(const Camera &camera);
         void Render(IDirect3DDevice9 *device, ID3DXMatrixStack *stack);
-        int GetFOV() { return m_fov; }
-        void SetFOV(int fov) { Dirty([&] { m_fov = fov; }, &m_fov); }
+        float GetFOV() { return m_fov; }
+        void SetFOV(float fov) { Dirty([&] { m_fov = fov; }, &m_fov); }
         nlohmann::json Save();
         void Load(const nlohmann::json &root);
 
     private:
-        int m_fov;
+        float m_fov;
     };
 }
